refactor(release): extract publication and env units out of release.ts
Step 1.1 of road-to-release-publication-integrity. The first refused attempt at
the placeholder guard died on check_source_size_budget: release.ts was 2,818 lines
against a 1,500-line ceiling, so any net growth in it is refused — including a
four-line version of the guard. This is the pure move that makes room.

THREE modules, and the shape was forced rather than chosen. A single cut does not
compile: the publication unit needs REMOTE, MAIN_BRANCH, CHANGELOG,
GH_PR_BODY_LIMIT, the Python-parity error classes and _cap_body from release.ts,
while release.ts needs die, run, git, gh and twenty more back. That is a cycle, and
a cycle is worse than the large file the split is fixing — so release_env.ts
carries the leaves both need.

  release.ts                2,818 -> 2,030   (net -788)
  release_publication.ts      727 (new)      the process/gh primitives plus
                                             everything that reaches GitHub
  release_env.ts              238 (new)      paths, branch names, GitHub body
                                             limits, three error classes,
                                             code-point helpers, _cap_body

Both new files sit under the ceiling and therefore contribute zero excess, which is
why the total falls by exactly what left release.ts: check_source_size_budget total
excess 19,363 -> 18,575, and the gate reported the ratchet loose, so the baseline is
lowered in the same change rather than leaving the gain silently reversible.

Re-export shape: release.ts already carried one trailing export block listing its
public surface, and the imports make every moved name resolve through it unchanged.
The six names tests import directly — _failed_check_names, _failed_checks_report,
_is_non_fast_forward, _is_tag_already_exists, _no_checks_action,
_required_contexts_from_rules — use `export ... from` rather than an import,
because they are not used in release.ts and an unused import would be dropped,
breaking the test path with nothing failing.

Pure move, no behaviour change: 133 release tests pass against the split,
unmodified.
