+++ src/config/gate-violation-baselines.json
@@ -23,10 +23,9 @@
             "note": "Lines under src/scripts/**/*.{ts,mts,mjs} that pass the retired `.agent-src.uncondensed` directory into `path.join(...)` \u2014 a filesystem path CONSTRUCTED against the tree ADR-051 abandoned, which is the exact shape of all 14 dead gates. Deliberately narrow so the number is signal, not noise: 236 raw mentions under src/scripts \u2192 147 string-literal lines \u2192 67 path-constructing lines when the pass was written, after exempting src/scripts/_lib/agent_src.ts (the shared resolver, whose job IS owning the constant \u2014 gates that route through it survived the move). Detectors holding the literal as data to match, strip, or print are not counted, and no prose, docs, tests or fixtures are read: that full-tree shape was measured at 44 files / 213 hits and rejected as the wrong lint (see the note in check_no_new_legacy_path.ts::find_offenders). Recorded at 58 rather than 67 because Phase 1's scan-root repairs were landing on this branch while the pass was being written and had already removed 9 \u2014 a baseline above the truth silently permits new violations. This count goes to ZERO as the remaining repairs land; every repair is a lowering commit. Lowered 56\u219251 by the road-to-renewal-foundation Phase 1 dead-root sweep (check_always_budget, check_portability, check_references, lint_framework_leakage, lint_agents_md). Lowered 51\u219249 by the gate-hardening adoption PR: skill_collision_clusters and skill_overlap were repointed at the live skills tree via the shared resolver \u2014 two gates that had been reading the retired container and scoring 0 of 288 skills. Lowered 49\u219248 by road-to-rule-delivery-integrity P2.2: score_skill_selection was repointed at the live skills tree via the shared resolver. It is the same shape as the two gates lowered before it \u2014 a scorer reading the retired container and scoring 0 of 288 skills \u2014 and here the consequence was sharper than a dead gate: the roadmap's pre-registered selection-accuracy criteria would have been compared against a baseline of silent zeros. Lowered 48\u219247 on 2026-08-13, and this one is a catch-up rather than a repair: the pass reported 47 against a baseline of 48, so a gain had already landed without the ratchet following it. Attributed to 455160160 (fix(evals): resolve the eval orchestrator against the live skills tree), the most recent repair in this branch's base that removes a path-constructing line \u2014 SKILLS_ROOT repointed through _lib/agent_src.ts::SRC_SKILLS(). A loose ratchet is not harmless: it silently permits one new violation to be added back without any gate noticing, which is the shape this entry exists to prevent."
         },
         "check_source_size_budget": {
-            "count": 19363,
-            "landed": "2026-08-22",
-            "note": "Lowered 19,395 -> 19,363 on 2026-08-22 by road-to-council-seat-selection, and the -32 is the residue of a LOWERING THROUGH EXTRACTION that had to pay for its own feature first. The branch adds the verified_at member stamp and the check_council_pin_staleness gate, whose 74 new lines landed in two files already far above the ceiling -- ai_council/config.ts (+63, at 2,272) and ai_council/clients.ts (+11, at 2,862). This gate refused it, correctly, and the debt could not be paid out of the additions themselves: of the 63 only a ~48-line validation block is a movable unit, and the 11 are a comment on a one-line constant. So the payment came from the cleanest pre-existing unit in the same file -- the already-labelled, already-self-contained 'Python-format helpers' section (114 lines) moved whole to ai_council/py_format.ts, which config.ts now imports. Net -32 past zero rather than exactly to it, because the section is a section and splitting it to hit a number would leave a worse file than moving it.",
-            "falsifier": "./scripts-run src/scripts/check_source_size_budget --quiet"
+            "count": 18575,
+            "landed": "2026-08-23",
+            "note": "Lowered 19,363 -> 18,575 on 2026-08-23 by road-to-release-publication-integrity step 1.1, and the -788 is a PURE MOVE that paid for nothing of its own. release.ts sat at 2,818 against the 1,500 ceiling, which is why the first two attempts at the placeholder guard were refused: any net growth in that file is refused, including a four-line version. The extraction is three modules rather than two, and the shape was forced rather than chosen -- a single cut left release.ts importing die/run/git back out of the publication unit while that unit imported REMOTE/MAIN_BRANCH back in, so release_env.ts holds the leaves (paths, branch names, GitHub body limits, the three Python-parity error classes, the code-point helpers and _cap_body) and both larger units import it. release.ts 2,818 -> 2,030; release_publication.ts 727 and release_env.ts 238, both under the ceiling and therefore contributing zero excess, which is what makes the total fall by exactly what left. Every moved public name is re-exported from release.ts, so no caller and no test import path changed: 133 release tests green against the split, unmodified."
         },
         "ci-parity:local-only": {
             "count": 165,
